+++ Libft/ft_memcmp.c
@@ -1,10 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcmp.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jbarbay <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/06 12:10:59 by jbarbay           #+#    #+#             */
+/*   Updated: 2023/09/06 12:11:03 by jbarbay          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
-	char *string1;
-	char *string2;
+	size_t	i;
+	char	*string1;
+	char	*string2;
 
 	string1 = (char *)s1;
 	string2 = (char *)s2;
