+++ Libft/ft_memmove.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memmove.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jbarbay <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/06 12:08:23 by jbarbay           #+#    #+#             */
+/*   Updated: 2023/09/06 12:08:26 by jbarbay          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void	*ft_memmove(void *dest, const void *src, int len)
+void	*ft_memmove(void *dest, const void *src, size_t len)
 {
-	char *source;
-	char *destination;
+	char	*source;
+	char	*destination;
 
 	source = (char *)src;
 	destination = (char *)dest;
